feat: add scroll with arrows
